idx-dibimbing-datawarehouse-main/requirements.txt
pandas==2.2.0
sqlalchemy==2.0.25
pymysql==1.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
cryptography==44.0.0
dbt-bigquery
dbt-postgres
dbt-core